Front e testes

+++ .docker/entrypoint.sh
@@ -3,8 +3,16 @@
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
 
 #chown -R www-data:www-data .
-cp .env.example .env
-cp .env.testing.example .env.testing
+npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
+
+cd backend
+if [ ! -f ".env" ]; then
+    cp .env.example .env
+fi
+if [ ! -f ".env.testing" ]; then
+    cp .env.testing.example .env.testing
+fi
 composer install
 php artisan key:generate
 php artisan migrate
